+++ src/Fable.Cli/CHANGELOG.md
@@ -10,6 +10,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 * [Python] - Print root module and module function comments (by @alfonsogarciacaro)
+* [Rust] Add support for module comments (by @ncave)
+* [Rust] Add support for null strings (by @ncave)
+
+### Fixed
+
 * [JS/TS] - Fix anonymous record printing (#4029) (by @alfonsogarciacaro)
 
 ## 5.0.0-alpha.9 - 2025-01-28